refactor(mcp): split run tool modules
